refactor: Refactor: split orbit-engine automation/vcs/pr.rs into focuse… [ORB-00119]
Refactor: split orbit-engine automation/vcs/pr.rs into focused submodules

Planned-By: codex
